Adding IE definitions

+++ firmware/openos/openwsn/02b-MAChigh/res.c
@@ -46,7 +46,7 @@ status information about several modules in the OpenWSN stack.
 bool debugPrint_myDAGrank() {
    uint16_t output=0;
    output = neighbors_getMyDAGrank();
-   openserial_printStatus(STATUS_DAGRANK,(uint16_t*)&output,sizeof(uint16_t));
+   openserial_printStatus(STATUS_DAGRANK,(uint8_t*)&output,sizeof(uint16_t));
    return TRUE;
 }
